feat: rename circ.rs to circle.rs
